refactored names of controller test package and test class names
